decouple app logic from window system using CRTP except for event type and remove all virtual functions
