+++ src/Utils/App.hpp
@@ -39,6 +39,8 @@
 #include "Utils/CappingProfiler.hpp"
 #include "Utils/Types.hpp"
 #include "Utils/DPI.hpp"
+#include "Utils/FileManager.hpp"
+#include "Utils/Scheduler.hpp"
 
 namespace VGG
 {
@@ -53,17 +55,56 @@ namespace VGG
  * It must be derived to use this class, and the singleton app
  * instance can be obtained by App::getIntance<DerivedApp>().
  */
+
+template<typename T>
 class App : public Uncopyable
 {
-public: // public data and types
-  struct SDLState
+// CRTP:
+// T must be implementes the following methods
+// initContext
+// makeContextCurrent
+// getProperty
+// swapBuffer
+// onInit
+// pollEvent
+// and dispatchEvent and dispatchGlobalEvent must be called properly to in pollEvent
+private:
+  inline void initContext(int w, int h, const std::string& title)
   {
-    SDL_Window* window{ nullptr };
-    SDL_GLContext glContext{ nullptr };
-  };
+    static_cast<T*>(this)->initContext(w, h, title);
+  }
+
+  inline void makeContextCurrent()
+  {
+    static_cast<T*>(this)->makeContextCurrent();
+  }
+
+  inline std::any getProperty(const std::string& name)
+  {
+    return static_cast<T*>(this)->getProperty(name);
+  }
+
+  inline void swapBuffer()
+  {
+    static_cast<T*>(this)->swapBuffer();
+  }
+
+  inline void onInit()
+  {
+    static_cast<T*>(this)->onInit();
+  }
+
+  inline void pollEvent()
+  {
+    static_cast<T*>(this)->pollEvent();
+  }
+
+public: // public data and types
   struct SkiaState
   {
-    sk_sp<const GrGLInterface> interface{ nullptr };
+    sk_sp<const GrGLInterface> interface {
+      nullptr
+    };
     sk_sp<GrDirectContext> grContext{ nullptr };
     sk_sp<SkSurface> surface{ nullptr };
 
@@ -77,6 +118,65 @@ public: // public data and types
     }
   };
 
+  struct Zoomer
+  {
+    static constexpr double minZoom = 0.01;
+    static constexpr double maxZoom = 10.0;
+    Vec2 offset{ 0.0, 0.0 };
+    double zoom{ 1.0 };
+    bool panning{ false };
+
+    void apply(SkCanvas* canvas)
+    {
+      ASSERT(canvas);
+      canvas->translate(offset.x, offset.y);
+      canvas->scale(zoom, zoom);
+    }
+
+    void restore(SkCanvas* canvas)
+    {
+      ASSERT(canvas);
+      canvas->scale(1. / zoom, 1. / zoom);
+      canvas->translate(-offset.x, -offset.y);
+    }
+
+    SDL_Event mapEvent(SDL_Event evt, double scaleFactor)
+    {
+      if (evt.type == SDL_MOUSEMOTION)
+      {
+        double x1 = evt.motion.x;
+        double y1 = evt.motion.y;
+        double x0 = x1 - evt.motion.xrel;
+        double y0 = y1 - evt.motion.yrel;
+        x1 = (x1 / scaleFactor - offset.x) / zoom;
+        y1 = (y1 / scaleFactor - offset.y) / zoom;
+        x0 = (x0 / scaleFactor - offset.x) / zoom;
+        y0 = (y0 / scaleFactor - offset.y) / zoom;
+        evt.motion.x = FLOAT2SINT(x1);
+        evt.motion.y = FLOAT2SINT(y1);
+        evt.motion.xrel = FLOAT2SINT(x1 - x0);
+        evt.motion.yrel = FLOAT2SINT(y1 - y0);
+      }
+      return evt;
+    }
+  };
+
+public:
+  inline void setOnFrameOnce(std::function<void()>&& cb)
+  {
+    Scheduler::setOnFrameOnce(std::move(cb));
+  }
+
+  inline void startRunMode()
+  {
+    EntityManager::setInteractionMode(EntityManager::InteractionMode::RUN);
+    InputManager::setMouseCursor(MouseEntity::CursorType::NORMAL);
+    if (auto container = EntityManager::getEntities())
+    {
+      container->setRunModeInteractions();
+    }
+  }
+
 protected: // protected members and static members
   static constexpr int N_MULTISAMPLE = 0;
   static constexpr int N_STENCILBITS = 8;
@@ -88,47 +188,8 @@ protected: // protected members and static members
   double m_pixelRatio;
   int m_nFrame;
   double m_timestamp;
-  SDLState m_sdlState;
   SkiaState m_skiaState;
-
-private: // private static methods
-  static inline void handle_sdl_error()
-  {
-    const char* err = SDL_GetError();
-    FAIL("SDL Error: %s", err);
-    SDL_ClearError();
-  }
-
-#ifdef __linux__
-  static double get_scale_factor()
-  {
-    static constexpr int NVARS = 4;
-    static const char* vars[NVARS] = {
-      "FORCE_SCALE",
-      "QT_SCALE_FACTOR",
-      "QT_SCREEN_SCALE_FACTOR",
-      "GDK_SCALE",
-    };
-    for (int i = 0; i < NVARS; i++)
-    {
-      const char* strVal = getenv(vars[i]);
-      if (strVal)
-      {
-        double val = atof(strVal);
-        if (val >= 1.0)
-        {
-          return val;
-        }
-      }
-    }
-    return 1.0;
-  }
-#else
-  static inline double get_scale_factor()
-  {
-    return 1.0;
-  }
-#endif
+  Zoomer m_zoomer;
 
   static bool init(App* app, int w, int h, const std::string& title)
   {
@@ -138,90 +199,10 @@ private: // private static methods
       return true;
     }
 
-    // init sdl
-    if (SDL_Init(SDL_INIT_VIDEO) != 0)
-    {
-      handle_sdl_error();
-      return false;
-    }
-    SDL_version compileVersion, linkVersion;
-    SDL_VERSION(&compileVersion);
-    SDL_GetVersion(&linkVersion);
-    INFO("SDL version %d.%d.%d (linked %d.%d.%d)",
-         compileVersion.major,
-         compileVersion.minor,
-         compileVersion.patch,
-         linkVersion.major,
-         linkVersion.minor,
-         linkVersion.patch);
+    app->initContext(w, h, title);
 
-    // setup opengl properties
-    SDL_GL_SetAttribute(SDL_GL_RED_SIZE, 8);
-    SDL_GL_SetAttribute(SDL_GL_GREEN_SIZE, 8);
-    SDL_GL_SetAttribute(SDL_GL_BLUE_SIZE, 8);
-    SDL_GL_SetAttribute(SDL_GL_ALPHA_SIZE, 8);
-    SDL_GL_SetAttribute(SDL_GL_BUFFER_SIZE, 1);
-    SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
-    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 8);
-    SDL_GL_SetAttribute(SDL_GL_STENCIL_SIZE, N_STENCILBITS);
-    SDL_GL_SetAttribute(SDL_GL_ACCELERATED_VISUAL, 1);
-    SDL_GL_SetAttribute(SDL_GL_MULTISAMPLEBUFFERS, 0);
-    SDL_GL_SetAttribute(SDL_GL_MULTISAMPLESAMPLES, N_MULTISAMPLE);
-#ifndef EMSCRIPTEN
-    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 3);
-    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 3);
-    SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE);
-#else
-    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 2);
-    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 0);
-    SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_ES);
-    SDL_SetHint(SDL_HINT_EMSCRIPTEN_KEYBOARD_ELEMENT, "#canvas");
-#endif
-
-    // create window
-    DPI::ScaleFactor = get_scale_factor();
-    int winWidth = w * DPI::ScaleFactor;
-    int winHeight = h * DPI::ScaleFactor;
-    SDL_Window* window =
-#ifndef EMSCRIPTEN
-      SDL_CreateWindow(title.c_str(),
-#else
-      SDL_CreateWindow(nullptr,
-#endif
-                       SDL_WINDOWPOS_CENTERED,
-                       SDL_WINDOWPOS_CENTERED,
-                       winWidth,
-                       winHeight,
-                       SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE | SDL_WINDOW_ALLOW_HIGHDPI);
-    if (!window)
-    {
-      handle_sdl_error();
-      return false;
-    }
-    app->m_width = w;
-    app->m_height = h;
-    app->m_sdlState.window = window;
-
-    // create context
-    SDL_GLContext glContext = SDL_GL_CreateContext(window);
-    if (!glContext)
-    {
-      handle_sdl_error();
-      return false;
-    }
-    app->m_sdlState.glContext = glContext;
-
-    // switch to this context
-    if (SDL_GL_MakeCurrent(window, glContext) != 0)
-    {
-      handle_sdl_error();
-      return false;
-    }
-    INFO("GL_VENDOR: %s", glGetString(GL_VENDOR));
-    INFO("GL_RENDERER: %s", glGetString(GL_RENDERER));
-    INFO("GL_VERSION: %s", glGetString(GL_VERSION));
-    INFO("GL_SHADING_LANGUAGE_VERSION: %s", glGetString(GL_SHADING_LANGUAGE_VERSION));
-
+    app->makeContextCurrent();
+    // Create Skia
     // get skia interface and make opengl context
     sk_sp<const GrGLInterface> interface = GrGLMakeNativeInterface();
     if (!interface)
@@ -235,23 +216,9 @@ private: // private static methods
       FAIL("Failed to make skia opengl context.");
       return false;
     }
+
     app->m_skiaState.interface = interface;
     app->m_skiaState.grContext = grContext;
-
-    // get device pixel ratio
-    int dw = 0, dh = 0;
-    int ww = 0, wh = 0;
-    SDL_GL_GetDrawableSize(window, &dw, &dh);
-    SDL_GetWindowSize(window, &ww, &wh);
-    DEBUG("Drawable size: %d %d", dw, dh);
-    DEBUG("Window size: %d %d", ww, wh);
-#ifdef __linux__
-    app->m_pixelRatio = DPI::ScaleFactor;
-#else
-    app->m_pixelRatio = (double)dw / ww;
-#endif
-    DEBUG("Scale factor: %.2lf", DPI::ScaleFactor);
-    DEBUG("Pixel ratio: %.2lf", app->m_pixelRatio);
 
     // get skia surface and canvas
     sk_sp<SkSurface> surface = app->setup_skia_surface(w, h);
@@ -268,8 +235,22 @@ private: // private static methods
       return false;
     }
 
-    // setup sdl configs
-    SDL_GL_SetSwapInterval(0);
+    // get necessary property about window and DPI
+    //
+    int dw = std::any_cast<int>(app->getProperty("viewport_w"));
+    int dh = std::any_cast<int>(app->getProperty("viewport_h"));
+    int ww = std::any_cast<int>(app->getProperty("window_w"));
+    int wh = std::any_cast<int>(app->getProperty("window_h"));
+
+    DEBUG("Drawable size: %d %d", dw, dh);
+    DEBUG("Window size: %d %d", ww, wh);
+#ifdef __linux__
+    app->m_pixelRatio = DPI::ScaleFactor;
+#else
+    app->m_pixelRatio = (double)dw / ww;
+#endif
+    DEBUG("Scale factor: %.2lf", DPI::ScaleFactor);
+    DEBUG("Pixel ratio: %.2lf", app->m_pixelRatio);
 
     // extra initialization
     app->onInit();
@@ -309,7 +290,18 @@ private: // private methods
                                                   &props);
   }
 
-  bool on_global_event(const SDL_Event& evt)
+protected: // protected methods
+  App()
+    : m_inited(false)
+    , m_shouldExit(false)
+    , m_width(0)
+    , m_height(0)
+    , m_pixelRatio(1.0)
+    , m_nFrame(0)
+  {
+  }
+
+  bool dispatchGlobalEvent(const SDL_Event& evt)
   {
     auto type = evt.type;
     auto key = evt.key.keysym.sym;
@@ -335,47 +327,107 @@ private: // private methods
     return onGlobalEvent(evt);
   }
 
-protected: // protected methods
-  App()
-    : m_inited(false)
-    , m_shouldExit(false)
-    , m_width(0)
-    , m_height(0)
-    , m_pixelRatio(1.0)
-    , m_nFrame(0)
+  void onFrame()
   {
-  }
 
-  virtual ~App()
-  {
-    if (m_inited && m_sdlState.glContext)
+    Scheduler::callOnFrameOnce();
+
+    if (SkCanvas* canvas = getCanvas())
     {
-      // NOTE The failure to delete GL context may be related to multi-thread and is hard
-      // to make it right. For simplicity, we can just safely ignore the deletion.
-      //
-      // SDL_GL_DeleteContext(m_sdlState.glContext);
+      m_zoomer.apply(canvas);
+      EntityManager::map([&](Entity& entity) { RenderSystem::drawEntity(canvas, entity); });
+      InputManager::draw(canvas);
+      m_zoomer.restore(canvas);
     }
-    if (m_inited && m_sdlState.window)
+
+    InputManager::onFrame();
+  }
+
+  void dispatchEvent(const SDL_Event& evt)
+  {
+    auto e = m_zoomer.mapEvent(evt, DPI::ScaleFactor);
+    InputManager::onEvent(e);
+  }
+
+  bool onGlobalEvent(const SDL_Event& evt)
+  {
+    auto type = evt.type;
+
+    if (type == SDL_WINDOWEVENT)
     {
-      SDL_DestroyWindow(m_sdlState.window);
+      if (evt.window.event == SDL_WINDOWEVENT_ENTER)
+      {
+        InputManager::setCursorVisibility(true);
+        return true;
+      }
+      else if (evt.window.event == SDL_WINDOWEVENT_LEAVE)
+      {
+        InputManager::setCursorVisibility(false);
+        return true;
+      }
     }
-    SDL_Quit();
-  }
 
-  virtual void onInit()
-  {
-  }
+    auto& panning = m_zoomer.panning;
+    if (!panning && type == SDL_MOUSEBUTTONDOWN &&
+        (SDL_GetKeyboardState(nullptr)[SDL_SCANCODE_SPACE]))
+    {
+      panning = true;
+      InputManager::setMouseCursor(MouseEntity::CursorType::MOVE);
+      return true;
+    }
+    else if (panning && type == SDL_MOUSEBUTTONUP)
+    {
+      panning = false;
+      InputManager::setMouseCursor(MouseEntity::CursorType::NORMAL);
+      return true;
+    }
+    else if (panning && type == SDL_MOUSEMOTION)
+    {
+      m_zoomer.offset.x += evt.motion.xrel / DPI::ScaleFactor;
+      m_zoomer.offset.y += evt.motion.yrel / DPI::ScaleFactor;
+      return true;
+    }
+    else if (type == SDL_MOUSEWHEEL && (SDL_GetModState() & KMOD_CTRL))
+    {
+      int mx, my;
+      SDL_GetMouseState(&mx, &my);
+      double dz = (evt.wheel.y > 0 ? 1.0 : -1.0) * 0.03;
+      double z2 = m_zoomer.zoom * (1 + dz);
+      if (z2 > 0.01 && z2 < 100)
+      {
+        m_zoomer.offset.x -= (mx / DPI::ScaleFactor - m_zoomer.offset.x) * dz;
+        m_zoomer.offset.y -= (my / DPI::ScaleFactor - m_zoomer.offset.y) * dz;
+        m_zoomer.zoom += m_zoomer.zoom * dz;
+        return true;
+      }
+      return false;
+    }
+    else if (type == SDL_KEYDOWN)
+    {
+      auto key = evt.key.keysym.sym;
+      auto mod = evt.key.keysym.mod;
 
-  virtual void onFrame()
-  {
-  }
+      if (key == SDLK_PAGEUP && (SDL_GetModState() & KMOD_CTRL))
+      {
+        FileManager::prevPage();
+        return true;
+      }
 
-  virtual void onEvent(const SDL_Event& evt)
-  {
-  }
+      if (key == SDLK_PAGEDOWN && (SDL_GetModState() & KMOD_CTRL))
+      {
+        FileManager::nextPage();
+        return true;
+      }
 
-  virtual bool onGlobalEvent(const SDL_Event& evt)
-  {
+#ifndef EMSCRIPTEN
+      if ((mod & KMOD_CTRL) && key == SDLK_q)
+      {
+        m_shouldExit = true;
+        return true;
+      }
+#endif
+    }
+
     return false;
   }
 
@@ -400,19 +452,8 @@ public: // public methods
     INFO("frame %d", m_nFrame++);
 #endif
 
+    pollEvent();
     // deal with events
-    SDL_Event evt;
-    while (SDL_PollEvent(&evt))
-    {
-      // process global events like shortcuts first
-      if (on_global_event(evt))
-      {
-        continue;
-      }
-
-      // process app events at last
-      onEvent(evt);
-    }
 
     // cap the frame rate
     auto profiler = CappingProfiler::getInstance();
@@ -436,18 +477,14 @@ public: // public methods
     canvas->flush();
     canvas->restore();
 
-    // display fps
-    SDL_SetWindowTitle(m_sdlState.window, profiler->fpsStr());
-
-    // swap buffer at last
-    SDL_GL_SwapWindow(m_sdlState.window);
+    swapBuffer();
   }
 
 public: // public static methods
   template<typename DerivedApp>
   static DerivedApp* getInstance(int w = 800, int h = 600, const std::string& title = "App")
   {
-    static_assert(std::is_base_of<App, DerivedApp>());
+    static_assert(std::is_base_of<App<DerivedApp>, DerivedApp>());
 
     static DerivedApp app;
