+++ src/runtime.cpp
@@ -16,186 +16,227 @@
  */
 #include <argparse/argparse.hpp>
 #include <algorithm>
+#include <filesystem>
 #include <nlohmann/json.hpp>
 
 #include "Entity/EntityManager.hpp"
 #include "Entity/InputManager.hpp"
 #include "Systems/RenderSystem.hpp"
 #include "Utils/App.hpp"
-#include "Utils/FileManager.hpp"
+#include "Utils/Utils.hpp"
 #include "Utils/Version.hpp"
-#include "Utils/Scheduler.hpp"
 
 using namespace VGG;
 
-class Runtime : public App
+class SDLRuntime : public App<SDLRuntime>
 {
-  struct Zoomer
+  struct SDLState
   {
-    static constexpr double minZoom = 0.01;
-    static constexpr double maxZoom = 10.0;
-    Vec2 offset{ 0.0, 0.0 };
-    double zoom{ 1.0 };
-    bool panning{ false };
-
-    void apply(SkCanvas* canvas)
-    {
-      ASSERT(canvas);
-      canvas->translate(offset.x, offset.y);
-      canvas->scale(zoom, zoom);
-    }
-
-    void restore(SkCanvas* canvas)
-    {
-      ASSERT(canvas);
-      canvas->scale(1. / zoom, 1. / zoom);
-      canvas->translate(-offset.x, -offset.y);
-    }
-
-    SDL_Event mapEvent(SDL_Event evt, double scaleFactor)
-    {
-      if (evt.type == SDL_MOUSEMOTION)
-      {
-        double x1 = evt.motion.x;
-        double y1 = evt.motion.y;
-        double x0 = x1 - evt.motion.xrel;
-        double y0 = y1 - evt.motion.yrel;
-        x1 = (x1 / scaleFactor - offset.x) / zoom;
-        y1 = (y1 / scaleFactor - offset.y) / zoom;
-        x0 = (x0 / scaleFactor - offset.x) / zoom;
-        y0 = (y0 / scaleFactor - offset.y) / zoom;
-        evt.motion.x = FLOAT2SINT(x1);
-        evt.motion.y = FLOAT2SINT(y1);
-        evt.motion.xrel = FLOAT2SINT(x1 - x0);
-        evt.motion.yrel = FLOAT2SINT(y1 - y0);
-      }
-      return evt;
-    }
+    SDL_Window* window{ nullptr };
+    SDL_GLContext glContext{ nullptr };
   };
-  Zoomer m_zoomer;
+  SDLState m_sdlState;
+  std::unordered_map<std::string, std::any> m_properties;
 
 public:
-  inline void setOnFrameOnce(std::function<void()>&& cb)
+  static inline void handle_sdl_error()
   {
-    Scheduler::setOnFrameOnce(std::move(cb));
+    const char* err = SDL_GetError();
+    FAIL("SDL Error: %s", err);
+    SDL_ClearError();
   }
 
-  inline void startRunMode()
+#ifdef __linux__
+  static double get_scale_factor()
   {
-    EntityManager::setInteractionMode(EntityManager::InteractionMode::RUN);
-    InputManager::setMouseCursor(MouseEntity::CursorType::NORMAL);
-    if (auto container = EntityManager::getEntities())
+    static constexpr int NVARS = 4;
+    static const char* vars[NVARS] = {
+      "FORCE_SCALE",
+      "QT_SCALE_FACTOR",
+      "QT_SCREEN_SCALE_FACTOR",
+      "GDK_SCALE",
+    };
+    for (int i = 0; i < NVARS; i++)
     {
-      container->setRunModeInteractions();
+      const char* strVal = getenv(vars[i]);
+      if (strVal)
+      {
+        double val = atof(strVal);
+        if (val >= 1.0)
+        {
+          return val;
+        }
+      }
     }
+    return 1.0;
+  }
+#else
+  static inline double get_scale_factor()
+  {
+    return 1.0;
+  }
+#endif
+  bool initContext(int w, int h, const std::string& title)
+  {
+    // init sdl
+    if (SDL_Init(SDL_INIT_VIDEO) != 0)
+    {
+      handle_sdl_error();
+      return false;
+    }
+    SDL_version compileVersion, linkVersion;
+    SDL_VERSION(&compileVersion);
+    SDL_GetVersion(&linkVersion);
+    INFO("SDL version %d.%d.%d (linked %d.%d.%d)",
+         compileVersion.major,
+         compileVersion.minor,
+         compileVersion.patch,
+         linkVersion.major,
+         linkVersion.minor,
+         linkVersion.patch);
+
+    // setup opengl properties
+    SDL_GL_SetAttribute(SDL_GL_RED_SIZE, 8);
+    SDL_GL_SetAttribute(SDL_GL_GREEN_SIZE, 8);
+    SDL_GL_SetAttribute(SDL_GL_BLUE_SIZE, 8);
+    SDL_GL_SetAttribute(SDL_GL_ALPHA_SIZE, 8);
+    SDL_GL_SetAttribute(SDL_GL_BUFFER_SIZE, 1);
+    SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
+    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 8);
+    SDL_GL_SetAttribute(SDL_GL_STENCIL_SIZE, N_STENCILBITS);
+    SDL_GL_SetAttribute(SDL_GL_ACCELERATED_VISUAL, 1);
+    SDL_GL_SetAttribute(SDL_GL_MULTISAMPLEBUFFERS, 0);
+    SDL_GL_SetAttribute(SDL_GL_MULTISAMPLESAMPLES, N_MULTISAMPLE);
+#ifndef EMSCRIPTEN
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 3);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 3);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE);
+#else
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 2);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 0);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_ES);
+    SDL_SetHint(SDL_HINT_EMSCRIPTEN_KEYBOARD_ELEMENT, "#canvas");
+#endif
+
+    // create window
+    DPI::ScaleFactor = get_scale_factor();
+    int winWidth = w * DPI::ScaleFactor;
+    int winHeight = h * DPI::ScaleFactor;
+    SDL_Window* window =
+#ifndef EMSCRIPTEN
+      SDL_CreateWindow(title.c_str(),
+#else
+      SDL_CreateWindow(nullptr,
+#endif
+                       SDL_WINDOWPOS_CENTERED,
+                       SDL_WINDOWPOS_CENTERED,
+                       winWidth,
+                       winHeight,
+                       SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE | SDL_WINDOW_ALLOW_HIGHDPI);
+    if (!window)
+    {
+      handle_sdl_error();
+      return false;
+    }
+    m_width = w;
+    m_height = h;
+    m_sdlState.window = window;
+    // create context
+    SDL_GLContext glContext = SDL_GL_CreateContext(window);
+    if (!glContext)
+    {
+      handle_sdl_error();
+      return false;
+    }
+
+    m_sdlState.glContext = glContext;
+
+    if (makeContextCurrent() == false)
+    {
+      return false;
+    }
+
+    // switch to this context
+    INFO("GL_VENDOR: %s", glGetString(GL_VENDOR));
+    INFO("GL_RENDERER: %s", glGetString(GL_RENDERER));
+    INFO("GL_VERSION: %s", glGetString(GL_VERSION));
+    INFO("GL_SHADING_LANGUAGE_VERSION: %s", glGetString(GL_SHADING_LANGUAGE_VERSION));
+
+    // get device pixel ratio
+    int dw = 0, dh = 0;
+    int ww = 0, wh = 0;
+    SDL_GL_GetDrawableSize(window, &dw, &dh);
+    SDL_GetWindowSize(window, &ww, &wh);
+    m_properties["viewport_w"] = dw;
+    m_properties["viewport_h"] = dh;
+    m_properties["window_w"] = ww;
+    m_properties["window_h"] = wh;
+    return true;
   }
 
-protected:
-  virtual void onInit() override
+  bool makeContextCurrent()
   {
+    WARN("makeContextCurrent");
+    if (SDL_GL_MakeCurrent(m_sdlState.window, m_sdlState.glContext) != 0)
+    {
+      handle_sdl_error();
+      return false;
+    }
+    return true;
+  }
+
+  std::any getProperty(const std::string& name)
+  {
+    return m_properties[name];
+  }
+
+  void onInit()
+  {
+    SDL_GL_SetSwapInterval(0);
     SDL_ShowCursor(SDL_DISABLE);
   }
 
-  virtual void onFrame() override
+  void pollEvent()
   {
-    Scheduler::callOnFrameOnce();
-
-    if (SkCanvas* canvas = getCanvas())
+    SDL_Event evt;
+    while (SDL_PollEvent(&evt))
     {
-      m_zoomer.apply(canvas);
-      EntityManager::map([&](Entity& entity) { RenderSystem::drawEntity(canvas, entity); });
-      InputManager::draw(canvas);
-      m_zoomer.restore(canvas);
-    }
+      // process global events like shortcuts first
+      if (dispatchGlobalEvent(evt))
+      {
+        continue;
+      }
 
-    InputManager::onFrame();
+      // process app events at last
+      dispatchEvent(evt);
+    }
   }
 
-  virtual void onEvent(const SDL_Event& evt) override
+  void swapBuffer()
   {
-    auto e = m_zoomer.mapEvent(evt, DPI::ScaleFactor);
-    InputManager::onEvent(e);
+
+    auto profiler = CappingProfiler::getInstance();
+    // display fps
+    SDL_SetWindowTitle(m_sdlState.window, profiler->fpsStr());
+
+    // swap buffer at last
+    SDL_GL_SwapWindow(m_sdlState.window);
   }
 
-  virtual bool onGlobalEvent(const SDL_Event& evt) override
+  ~SDLRuntime()
   {
-    auto type = evt.type;
-
-    if (type == SDL_WINDOWEVENT)
+    if (m_inited && m_sdlState.glContext)
     {
-      if (evt.window.event == SDL_WINDOWEVENT_ENTER)
-      {
-        InputManager::setCursorVisibility(true);
-        return true;
-      }
-      else if (evt.window.event == SDL_WINDOWEVENT_LEAVE)
-      {
-        InputManager::setCursorVisibility(false);
-        return true;
-      }
+      // NOTE The failure to delete GL context may be related to multi-thread and is hard
+      // to make it right. For simplicity, we can just safely ignore the deletion.
+      //
+      // SDL_GL_DeleteContext(m_sdlState.glContext);
     }
-
-    auto& panning = m_zoomer.panning;
-    if (!panning && type == SDL_MOUSEBUTTONDOWN && (SDL_GetKeyboardState(nullptr)[SDL_SCANCODE_SPACE]))
+    if (m_inited && m_sdlState.window)
     {
-      panning = true;
-      InputManager::setMouseCursor(MouseEntity::CursorType::MOVE);
-      return true;
+      SDL_DestroyWindow(m_sdlState.window);
     }
-    else if (panning && type == SDL_MOUSEBUTTONUP)
-    {
-      panning = false;
-      InputManager::setMouseCursor(MouseEntity::CursorType::NORMAL);
-      return true;
-    }
-    else if (panning && type == SDL_MOUSEMOTION)
-    {
-      m_zoomer.offset.x += evt.motion.xrel / DPI::ScaleFactor;
-      m_zoomer.offset.y += evt.motion.yrel / DPI::ScaleFactor;
-      return true;
-    }
-    else if (type == SDL_MOUSEWHEEL && (SDL_GetModState() & KMOD_CTRL))
-    {
-      int mx, my;
-      SDL_GetMouseState(&mx, &my);
-      double dz = (evt.wheel.y > 0 ? 1.0 : -1.0) * 0.03;
-      double z2 = m_zoomer.zoom * (1 + dz);
-      if (z2 > 0.01 && z2 < 100)
-      {
-        m_zoomer.offset.x -= (mx / DPI::ScaleFactor - m_zoomer.offset.x) * dz;
-        m_zoomer.offset.y -= (my / DPI::ScaleFactor - m_zoomer.offset.y) * dz;
-        m_zoomer.zoom += m_zoomer.zoom * dz;
-        return true;
-      }
-      return false;
-    }
-    else if (type == SDL_KEYDOWN)
-    {
-      auto key = evt.key.keysym.sym;
-      auto mod = evt.key.keysym.mod;
-
-      if (key == SDLK_PAGEUP && (SDL_GetModState() & KMOD_CTRL))
-      {
-        FileManager::prevPage();
-        return true;
-      }
-
-      if (key == SDLK_PAGEDOWN && (SDL_GetModState() & KMOD_CTRL))
-      {
-        FileManager::nextPage();
-        return true;
-      }
-
-#ifndef EMSCRIPTEN
-      if ((mod & KMOD_CTRL) && key == SDLK_q)
-      {
-        m_shouldExit = true;
-        return true;
-      }
-#endif
-    }
-
-    return false;
+    SDL_Quit();
   }
 };
 
@@ -277,7 +318,7 @@ int main(int argc, char** argv)
     }
   }
 
-  Runtime* app = App::getInstance<Runtime>(1200, 800, "VGG");
+  SDLRuntime* app = App<SDLRuntime>::getInstance<SDLRuntime>(1200, 800, "VGG");
   ASSERT(app);
   if (auto fm = FileManager::getInstance(); fm && fm->fileCount() < 1)
   {
